+++ CHANGELOG.md
@@ -4,11 +4,13 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 
 ## ## [x.x.x] - {yyyy-mm-dd}
 ### Added
+* Logging outbound email sends to database, including status and forensics, to troubleshoot silent failures
 ### Changed
 * Users can download previous versions of a Manuscript
 ### Deprecated
 ### Removed
 ### Fixed
+* Users can now select and copy text from invitation letters when there is more than one invitation.
 ### Security
 
 ## ## [1.34.0] - {2017-01-06}