+++ CHANGELOG.md
@@ -6,6 +6,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ### Added
 * Logging outbound email sends to database, including status and forensics, to troubleshoot silent failures
 ### Changed
+* Users can download previous versions of a Manuscript
 ### Deprecated
 ### Removed
 ### Fixed